Move commands to bin/glusterd2/commands
