+++ bin/glusterd2/commands/peers/addpeer.go
