+++ bin/glusterd2/commands/peers/deletepeer.go
