+++ bin/glusterd2/commands/peers/getpeer.go
