+++ bin/glusterd2/commands/peers/getpeers.go
