+++ bin/glusterd2/commands/peers/peer-status.go
