+++ bin/glusterd2/commands/version/version.go
