+++ bin/glusterd2/commands/volumes/volume-delete.go
