+++ bin/glusterd2/commands/volumes/volume-stop.go
